refactor: slim down the entry script
